Moved pygame-related code up to the Apex class.
